making progress on enlil 2.8 reader update.  Have one more small bug to get the hypercube extraction working correctly.... need to include whole-extents of the file when configuring...

+++ ghost-kit/enlil/src/enlil3dfile.h
@@ -77,7 +77,7 @@ protected:  //DATA
 
     //Data Keeping
     double _MJD;
-    qint64 _TIME;
+    double _TIME;
     int _dims[3];
     double _enlil_version;
     int _extents[6];
